refactor: extract control capability calculations
